@@ -12,9 +12,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.aggregation.Aggregation;
@@ -23,44 +23,16 @@ import org.springframework.data.mongodb.core.aggregation.GroupOperation;
 import org.springframework.test.context.TestPropertySource;
 
 import com.baeldung.logging.model.Book;
-import com.mongodb.client.MongoClients;
-
-import de.flapdoodle.embed.mongo.MongodExecutable;
-import de.flapdoodle.embed.mongo.MongodStarter;
-import de.flapdoodle.embed.mongo.config.ImmutableMongodConfig;
-import de.flapdoodle.embed.mongo.config.MongodConfig;
-import de.flapdoodle.embed.mongo.config.Net;
-import de.flapdoodle.embed.mongo.distribution.Version;
-import de.flapdoodle.embed.process.runtime.Network;
 
 @SpringBootTest(classes = SpringBootLoggingApplication.class)
-@TestPropertySource(properties = {"logging.level.org.springframework.data.mongodb.core.MongoTemplate=WARN"}, value = "/embedded.properties")
+@TestPropertySource(properties = { "logging.level.org.springframework.data.mongodb.core.MongoTemplate=WARN" }, value = "/embedded.properties")
 public class LoggingUnitTest {
 
-    private static final String CONNECTION_STRING = "mongodb://%s:%d";
-
-    private MongodExecutable mongodExecutable;
-    private MongoTemplate mongoTemplate;
-
-    @AfterEach
-    void clean() {
-        mongodExecutable.stop();
-    }
+    private @Autowired MongoTemplate mongoTemplate;
 
     @BeforeEach
     void setup() throws Exception {
-        String ip = "localhost";
-        int port = Network.freeServerPort(Network.getLocalHost());
-
-        ImmutableMongodConfig mongodConfig = MongodConfig.builder()
-                .version(Version.Main.PRODUCTION)
-                .net(new Net(ip, port, Network.localhostIsIPv6()))
-                .build();
-
-        MongodStarter starter = MongodStarter.getDefaultInstance();
-        mongodExecutable = starter.prepare(mongodConfig);
-        mongodExecutable.start();
-        mongoTemplate = new MongoTemplate(MongoClients.create(String.format(CONNECTION_STRING, ip, port)), "test");
+        mongoTemplate.dropCollection("book");
     }
 
     @Test
@@ -88,7 +60,7 @@ public class LoggingUnitTest {
         mongoTemplate.updateFirst(query(where("bookName").is("Book")), update("authorName", authorNameUpdate), Book.class);
 
         assertThat(mongoTemplate.findById(book.getId(), Book.class)).extracting(Book::getAuthorName)
-                .isEqualTo(authorNameUpdate);
+            .isEqualTo(authorNameUpdate);
     }
 
     @Test
@@ -104,7 +76,7 @@ public class LoggingUnitTest {
         mongoTemplate.insert(Arrays.asList(book, book1), Book.class);
 
         assertThat(mongoTemplate.findAll(Book.class)
-                .size()).isEqualTo(2);
+            .size()).isEqualTo(2);
     }
 
     @Test
@@ -118,7 +90,7 @@ public class LoggingUnitTest {
         mongoTemplate.remove(book);
 
         assertThat(mongoTemplate.findAll(Book.class)
-                .size()).isEqualTo(0);
+            .size()).isEqualTo(0);
     }
 
     @Test
@@ -138,21 +110,21 @@ public class LoggingUnitTest {
         mongoTemplate.insert(Arrays.asList(book, book1, book2), Book.class);
 
         GroupOperation groupByAuthor = group("authorName").count()
-                .as("authCount");
+            .as("authCount");
 
         Aggregation aggregation = newAggregation(groupByAuthor);
 
         AggregationResults<GroupByAuthor> aggregationResults = mongoTemplate.aggregate(aggregation, "book", GroupByAuthor.class);
 
         List<GroupByAuthor> groupByAuthorList = StreamSupport.stream(aggregationResults.spliterator(), false)
-                .collect(Collectors.toList());
+            .collect(Collectors.toList());
 
         assertThat(groupByAuthorList.stream()
-                .filter(l -> l.getAuthorName()
-                        .equals("Author"))
-                .findFirst()
-                .orElse(null)).extracting(GroupByAuthor::getAuthCount)
-                .isEqualTo(3);
+            .filter(l -> l.getAuthorName()
+                .equals("Author"))
+            .findFirst()
+            .orElse(null)).extracting(GroupByAuthor::getAuthCount)
+            .isEqualTo(3);
     }
 
 }
