@@ -1,24 +1,30 @@
 package com.baeldung.hibernate;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.io.IOException;
+
+import org.hibernate.Session;
+import org.hibernate.Transaction;
+import org.hibernate.id.IdentifierGenerationException;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.hibernate.Transaction;
-import java.io.IOException;
-import org.hibernate.Session;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.baeldung.hibernate.pojo.Product;
+import com.baeldung.hibernate.pojo.BaseballPlayer;
 import com.baeldung.hibernate.pojo.Course;
 import com.baeldung.hibernate.pojo.OrderEntry;
 import com.baeldung.hibernate.pojo.OrderEntryIdClass;
 import com.baeldung.hibernate.pojo.OrderEntryPK;
+import com.baeldung.hibernate.pojo.Product;
 import com.baeldung.hibernate.pojo.Student;
+import com.baeldung.hibernate.pojo.TennisPlayer;
 import com.baeldung.hibernate.pojo.User;
 import com.baeldung.hibernate.pojo.UserProfile;
 
 public class IdentifiersIntegrationTest {
+
     private Session session;
 
     private Transaction transaction;
@@ -37,12 +43,32 @@ public class IdentifiersIntegrationTest {
     }
 
     @Test
-    public void whenSaveSimpleIdEntities_thenOk() {
+    public void whenSavingTennisPlayerWithoutAnId_thenSavingEntityOk() {
+        TennisPlayer tennisPlayer = new TennisPlayer("Tom");
+        session.save(tennisPlayer);
+        assertThat(tennisPlayer.getPlayerId()).isEqualTo(0L);
+    }
+
+    @Test
+    public void whenSavingBaseballPlayerWithoutAnId_thenSavingEntityFails() {
+        BaseballPlayer baseballPlayer = new BaseballPlayer("Jerry");
+        assertThatThrownBy(() -> session.save(baseballPlayer)).isInstanceOf(IdentifierGenerationException.class)
+            .hasMessageContaining("ids for this class must be manually assigned before calling save()");
+    }
+    @Test
+    public void whenSavingBaseballPlayerWithAManualId_thenSavingEntityOK() {
+        BaseballPlayer baseballPlayer = new BaseballPlayer("Jerry");
+        baseballPlayer.setPlayerId(42L);
+        session.save(baseballPlayer);
+    }
+
+    @Test
+    public void whenSaveSequenceIdEntities_thenOk() {
         Student student = new Student();
         session.save(student);
         User user = new User();
         session.save(user);
-       
+
         assertThat(student.getStudentId()).isEqualTo(1L);
         assertThat(user.getUserId()).isEqualTo(4L);
 
