@@ -2,16 +2,12 @@ package com.baeldung.rest.karate;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
-import com.intuit.karate.junit4.Karate;
+import com.intuit.karate.junit5.Karate;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.AfterAll;
-import org.junit.runner.RunWith;
-
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 
-@RunWith(Karate.class)
 public class KarateIntegrationTest {
 
     private static final int PORT_NUMBER = 8097;
@@ -41,6 +37,11 @@ public class KarateIntegrationTest {
     @AfterAll
     public static void tearDown() {
         wireMockServer.stop();
+    }
+
+    @Karate.Test
+    Karate testAll() {
+        return Karate.run().relativeTo(getClass());
     }
 
 }
